complete previous rename
