+++ src/App.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import './App.css';
 
-import Game from './components/Screen';
+import Game from './components/Game';
 import Welcome from './components/Welcome';
 
 function App() {
